+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## [0.1.1] 19-09-2022
+
+- Add reactable.read and reactable.write to change the value silently.
+- Add Scoped view to wrap a widget with scope.
+
 ## [0.1.0] 16-07-2022
 
 Test release.